refactor(server): extract sessions mutate handlers (#593)
Extract the 10 mutate handlers (api_active_session, api_create_session,
api_close_session, api_session_config, api_recreate_session,
api_spawn_sibling, api_fork_session, api_session_broadcast,
api_restart_service, api_refresh_sessions) from server.py into
agentwire/routes/sessions_admin.py as SessionsAdminRoutesMixin,
completing the #593 sessions split (PR-B of 2).

Shared state and helpers (active_sessions, _get_sessions_data,
_get_session_config, _get_session_cwd, _deliver_first_message,
_fetch_remote_machine_sessions, yaml read/write helpers) stay on the
base server. Session is imported lazily from ..server inside
api_session_broadcast to avoid a circular import.

Closes #593

Built by dotdev.dev
